fix: Fix circular import by KernelLiveStatValues
